requirements.txt
flask
python-valve